+++ app/api/api/serializers.py
@@ -26,7 +26,7 @@ from mapping.models import (
 )
 from users.serializers import UserSerializer
 from mapping.permissions import has_editorship, is_admin, is_az_function_user
-from shared.services.rules_export import analyse_concepts
+from services.rules_export import analyse_concepts
 from config.settings import DATA_UPLOAD_MAX_MEMORY_SIZE
