+++ app/api/mapping/management/commands/generate_rules_json.py
@@ -2,7 +2,7 @@ import json
 
 from django.core.management.base import BaseCommand
 from mapping.models import MappingRule
-from shared.services.rules_export import get_mapping_rules_json, make_dag
+from services.rules_export import get_mapping_rules_json, make_dag
 
 
 class Command(BaseCommand):
